Refactor for running bridge with eth_portal.bridge
