+++ eth_portal/bridge/__main__.py
@@ -1,0 +1,6 @@
+from .run import launch_bridge
+
+try:
+    launch_bridge()
+except KeyboardInterrupt:
+    print("Clean exit of bridge launcher")
