+++ newsfragments/23.feature.rst
@@ -1,0 +1,1 @@
+Now launch the bridge with ``python -m eth_portal.bridge``.
